web library

+++ src/components/axios.js
@@ -3,7 +3,7 @@ import axios from "axios";
 import { Link } from "react-router-dom";
 
 function App() {
-  const url = "http://localhost:3000/books";
+  const url = "http://localhost:8080/books";
 
   const [data, setData] = useState({ data: [] });
   useEffect(() => {
@@ -12,32 +12,49 @@ function App() {
 
   function onDelete(id) {
     axios
-      .delete(`http://localhost:3000/books/${id}`)
+      .delete(`http://localhost:8080/books/${id}`)
       .then(alert("Book deleted!"));
     window.location.reload(false);
   }
+
+  const role = localStorage.getItem("jwtRole");
 
   const renderTable = () => {
     return data.data.map(book => {
       return (
         <tr key={book.id}>
+          <td>{book.id}</td>
           <td>{book.title}</td>
           <td>{book.author}</td>
           <td>{book.page}</td>
           <td>{book.language}</td>
           <td>{book.publisher_id}</td>
           <td>
-            <Link to={"/updatebook/" + book.id}>
-              <button className="button btn-warning btn-sm btn-block">
-                Edit
-              </button>{" "}
-            </Link>
-            <button
-              className="button btn-danger btn-sm btn-block"
-              onClick={() => onDelete(book.id)}
-            >
-              Delete
-            </button>
+            {(() => {
+              if (role == "ADMIN") {
+                return (
+                  <>
+                    <Link to={"/updatebook/" + book.id}>
+                      <button className="button btn-warning btn-sm btn-block">
+                        Edit
+                      </button>
+                    </Link>
+                    <button
+                      className="button btn-danger btn-sm btn-block"
+                      onClick={() => onDelete(book.id)}
+                    >
+                      Delete
+                    </button>
+                  </>
+                );
+              } else {
+                return (
+                  <button className="button btn-warning btn-sm btn-block">
+                    Rent
+                  </button>
+                );
+              }
+            })()}
           </td>
         </tr>
       );
@@ -46,10 +63,12 @@ function App() {
 
   return (
     <div className="table-container">
-      <h5>Users</h5>
-      <table class="table table-bordered">
+      <h5>Books</h5>
+      <table className="table table-bordered">
         <thead>
           <tr>
+            <th>ID</th>
+
             <th>Title</th>
             <th>Author</th>
             <th>Page</th>
